Fix in generate integrations docs script (#16108)

+++ integrations/gen_docs_integrations.py
@@ -24,7 +24,7 @@ def cleanup():
             if "integrations" in str(element):
                 shutil.rmtree(element)
         for element in Path("integrations/cloud-notifications").glob('**/*/'):
-            if "integrations" in str(element):
+            if "integrations" in str(element) and not "metadata.yaml" in str(element):
                 shutil.rmtree(element)
